pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "claude-meet-mcp"
version = "1.0.0"
description = "AI-powered meeting scheduler using Claude and Google Calendar. Stop playing email tennis."
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    {name = "Shaunak Inamdar", email = "inamdarshaunak01@gmail.com"}
]
maintainers = [
    {name = "Shaunak Inamdar", email = "inamdarshaunak01@gmail.com"}
]
keywords = [
    "calendar",
    "scheduling",
    "claude",
    "anthropic",
    "ai",
    "google-calendar",
    "mcp",
    "model-context-protocol",
    "meeting",
    "productivity",
    "automation"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Scheduling",
    "Topic :: Communications :: Conferencing",
    "Topic :: Utilities",
]

dependencies = [
    "anthropic>=0.40.0",
    "google-api-python-client>=2.100.0",
    "google-auth-httplib2>=0.1.1",
    "google-auth-oauthlib>=1.1.0",
    "click>=8.1.7",
    "python-dateutil>=2.8.2",
    "pytz>=2023.3",
    "python-dotenv>=1.0.0",
    "mcp>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "build",
    "twine",
]

[project.scripts]
claude-meet = "claude_meet.cli:main"
claude-meet-mcp = "claude_meet.mcp_server:main"

[project.urls]
Homepage = "https://github.com/ShaunakInamdar/claude-meet-mcp"
Documentation = "https://github.com/ShaunakInamdar/claude-meet-mcp#readme"
Repository = "https://github.com/ShaunakInamdar/claude-meet-mcp"
Issues = "https://github.com/ShaunakInamdar/claude-meet-mcp/issues"
Changelog = "https://github.com/ShaunakInamdar/claude-meet-mcp/releases"

[tool.setuptools]
packages = ["claude_meet"]
include-package-data = true

[tool.setuptools.package-data]
claude_meet = ["py.typed"]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "C901",   # too complex
    "B904",   # raise without from inside except
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B011"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
